chore: make map-to-vfiles operate on a single doc
